requirements.txt
python-dotenv
python-telegram-bot==13.15
python-binance
